requirements.txt
python-dotenv==1.0.0
openai==1.6.1
tqdm==4.66.1
openpyxl==3.1.2